[add]password setting login, logout button
