+++ Assets/TitleScenes/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 982ba65442526694882a9408651da14b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
